Split uplink / downlink data into separate files.
