+++ internal/loraserver/loraserver.go
@@ -1,11 +1,8 @@
 package loraserver
 
 import (
-	"errors"
 	"fmt"
-	"strings"
 	"sync"
-	"time"
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/brocaar/loraserver/models"
@@ -100,343 +97,4 @@ func handleRXPacket(ctx Context, rxPacket models.RXPacket) error {
 	default:
 		return fmt.Errorf("unknown MType: %v", rxPacket.PHYPayload.MHDR.MType)
 	}
-}
-
-func validateAndCollectDataUpRXPacket(ctx Context, rxPacket models.RXPacket) error {
-	// MACPayload must be of type *lorawan.MACPayload
-	macPL, ok := rxPacket.PHYPayload.MACPayload.(*lorawan.MACPayload)
-	if !ok {
-		return fmt.Errorf("expected *lorawan.MACPayload, got: %T", rxPacket.PHYPayload.MACPayload)
-	}
-
-	// get the session data
-	ns, err := getNodeSession(ctx.RedisPool, macPL.FHDR.DevAddr)
-	if err != nil {
-		return err
-	}
-
-	// validate and get the full int32 FCnt
-	fullFCnt, ok := validateAndGetFullFCntUp(ns, macPL.FHDR.FCnt)
-	if !ok {
-		log.WithFields(log.Fields{
-			"dev_addr":    macPL.FHDR.DevAddr,
-			"dev_eui":     ns.DevEUI,
-			"packet_fcnt": macPL.FHDR.FCnt,
-			"server_fcnt": ns.FCntUp,
-		}).Warning("invalid FCnt")
-		return errors.New("invalid FCnt or too many dropped frames")
-	}
-	macPL.FHDR.FCnt = fullFCnt
-
-	// validate MIC
-	micOK, err := rxPacket.PHYPayload.ValidateMIC(ns.NwkSKey)
-	if err != nil {
-		return fmt.Errorf("validate MIC error: %s", err)
-	}
-	if !micOK {
-		return errors.New("invalid MIC")
-	}
-
-	if macPL.FPort != nil {
-		if *macPL.FPort == 0 {
-			// decrypt FRMPayload with NwkSKey when FPort == 0
-			if err := rxPacket.PHYPayload.DecryptFRMPayload(ns.NwkSKey); err != nil {
-				return fmt.Errorf("decrypt FRMPayload error: %s", err)
-			}
-		} else {
-			if err := rxPacket.PHYPayload.DecryptFRMPayload(ns.AppSKey); err != nil {
-				return fmt.Errorf("decrypt FRMPayload error: %s", err)
-			}
-		}
-	}
-
-	return collectAndCallOnce(ctx.RedisPool, rxPacket, func(rxPackets RXPackets) error {
-		return handleCollectedDataUpPackets(ctx, rxPackets)
-	})
-}
-
-func handleCollectedDataUpPackets(ctx Context, rxPackets RXPackets) error {
-	if len(rxPackets) == 0 {
-		return errors.New("packet collector returned 0 packets")
-	}
-	rxPacket := rxPackets[0]
-
-	var macs []string
-	for _, p := range rxPackets {
-		macs = append(macs, p.RXInfo.MAC.String())
-	}
-
-	macPL, ok := rxPacket.PHYPayload.MACPayload.(*lorawan.MACPayload)
-	if !ok {
-		return fmt.Errorf("expected *lorawan.MACPayload, got: %T", rxPacket.PHYPayload.MACPayload)
-	}
-
-	ns, err := getNodeSession(ctx.RedisPool, macPL.FHDR.DevAddr)
-	if err != nil {
-		return err
-	}
-
-	log.WithFields(log.Fields{
-		"dev_eui":  ns.DevEUI,
-		"gw_count": len(rxPackets),
-		"gw_macs":  strings.Join(macs, ", "),
-		"mtype":    rxPackets[0].PHYPayload.MHDR.MType,
-	}).Info("packet(s) collected")
-
-	// send rx info notification to be used by the network-controller
-	if err = sendRXInfoPayload(ctx, ns.AppEUI, ns.DevEUI, rxPackets); err != nil {
-		return fmt.Errorf("send rx info notification error: %s", err)
-	}
-
-	// handle FOpts mac commands (if any)
-	if len(macPL.FHDR.FOpts) > 0 {
-		if err := handleUplinkMACCommands(ctx, ns.AppEUI, ns.DevEUI, false, macPL.FHDR.FOpts); err != nil {
-			return fmt.Errorf("handle FOpts mac commands error: %s", err)
-		}
-	}
-
-	if macPL.FPort != nil {
-		if *macPL.FPort == 0 {
-			if len(macPL.FRMPayload) == 0 {
-				return errors.New("expected mac commands, but FRMPayload is empty (FPort=0)")
-			}
-			var commands []lorawan.MACCommand
-			for _, pl := range macPL.FRMPayload {
-				cmd, ok := pl.(*lorawan.MACCommand)
-				if !ok {
-					return fmt.Errorf("expected MACPayload, but got %T", macPL.FRMPayload)
-				}
-				commands = append(commands, *cmd)
-			}
-			if err := handleUplinkMACCommands(ctx, ns.AppEUI, ns.DevEUI, true, commands); err != nil {
-				return fmt.Errorf("handle FRMPayload mac commands error: %s", err)
-			}
-		} else {
-			var data []byte
-
-			// it is possible that the FRMPayload is empty, in this case only
-			// the FPort will be send
-			if len(macPL.FRMPayload) == 1 {
-				dataPL, ok := macPL.FRMPayload[0].(*lorawan.DataPayload)
-				if !ok {
-					return errors.New("FRMPayload must be of type *lorawan.DataPayload")
-				}
-				data = dataPL.Bytes
-			}
-
-			err = ctx.Application.SendRXPayload(ns.AppEUI, ns.DevEUI, models.RXPayload{
-				DevEUI:       ns.DevEUI,
-				GatewayCount: len(rxPackets),
-				FPort:        *macPL.FPort,
-				RSSI:         rxPacket.RXInfo.RSSI,
-				Data:         data,
-			})
-			if err != nil {
-				return fmt.Errorf("send rx payload to application error: %s", err)
-			}
-		}
-	}
-
-	// sync counter with that of the device
-	ns.FCntUp = macPL.FHDR.FCnt
-	if err := saveNodeSession(ctx.RedisPool, ns); err != nil {
-		return err
-	}
-
-	// handle downlink (ACK)
-	time.Sleep(CollectDataDownWait)
-	if err := handleDataDownReply(ctx, rxPacket, ns); err != nil {
-		return fmt.Errorf("handling downlink data for node %s failed: %s", ns.DevEUI, err)
-	}
-
-	return nil
-}
-
-func handleUplinkMACCommands(ctx Context, appEUI, devEUI lorawan.EUI64, frmPayload bool, commands []lorawan.MACCommand) error {
-	for _, cmd := range commands {
-		log.WithFields(log.Fields{
-			"dev_eui":     devEUI,
-			"cid":         cmd.CID,
-			"frm_payload": frmPayload,
-		}).Info("mac command received")
-
-		b, err := cmd.MarshalBinary()
-		if err != nil {
-			return fmt.Errorf("binary marshal mac command error: %s", err)
-		}
-
-		err = ctx.Controller.SendMACPayload(appEUI, devEUI, models.MACPayload{
-			DevEUI:     devEUI,
-			FRMPayload: frmPayload,
-			MACCommand: b,
-		})
-		if err != nil {
-			return fmt.Errorf("send mac payload error: %s", err)
-		}
-	}
-	return nil
-}
-
-func handleDataDownReply(ctx Context, rxPacket models.RXPacket, ns models.NodeSession) error {
-	macPL, ok := rxPacket.PHYPayload.MACPayload.(*lorawan.MACPayload)
-	if !ok {
-		return fmt.Errorf("expected *lorawan.MACPayload, got: %T", rxPacket.PHYPayload.MACPayload)
-	}
-
-	// the last payload was received by the node
-	if macPL.FHDR.FCtrl.ACK {
-		txPayload, err := clearInProcessTXPayload(ctx.RedisPool, ns.DevEUI)
-		if err != nil {
-			return err
-		}
-		ns.FCntDown++
-		if err = saveNodeSession(ctx.RedisPool, ns); err != nil {
-			return err
-		}
-		if txPayload != nil {
-			err = ctx.Application.SendNotification(ns.AppEUI, ns.DevEUI, models.ACKNotificationType, models.ACKNotification{
-				Reference: txPayload.Reference,
-				DevEUI:    ns.DevEUI,
-			})
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	// check if there are payloads pending in the queue
-	txPayload, remaining, err := getTXPayloadAndRemainingFromQueue(ctx.RedisPool, ns.DevEUI)
-	if err != nil {
-		return err
-	}
-
-	// nothing pending in the queue and no need to ACK RXPacket
-	if rxPacket.PHYPayload.MHDR.MType != lorawan.ConfirmedDataUp && txPayload == nil {
-		return nil
-	}
-
-	// get TX DR
-	uplinkDR, err := Band.GetDataRate(rxPacket.RXInfo.DataRate)
-	if err != nil {
-		return err
-	}
-	// get TX channel
-	uplinkChannel, err := Band.GetChannel(rxPacket.RXInfo.Frequency, uplinkDR)
-	if err != nil {
-		return err
-	}
-	// get RX1 channel
-	rx1Channel := Band.GetRX1Channel(uplinkChannel)
-	// get RX1 DR
-	rx1DR, err := Band.GetRX1DataRateForOffset(uplinkDR, int(ns.RX1DROffset))
-	if err != nil {
-		return err
-	}
-	// get rx delay
-	rxDelay := Band.ReceiveDelay1
-	if ns.RXDelay > 0 {
-		rxDelay = time.Duration(ns.RXDelay) * time.Second
-	}
-
-	phy := lorawan.PHYPayload{
-		MHDR: lorawan.MHDR{
-			MType: lorawan.UnconfirmedDataDown,
-			Major: lorawan.LoRaWANR1,
-		},
-	}
-	macPL = &lorawan.MACPayload{
-		FHDR: lorawan.FHDR{
-			DevAddr: ns.DevAddr,
-			FCtrl: lorawan.FCtrl{
-				ACK: rxPacket.PHYPayload.MHDR.MType == lorawan.ConfirmedDataUp, // set ACK to true when received packet needs an ACK
-			},
-			FCnt: ns.FCntDown,
-		},
-	}
-	phy.MACPayload = macPL
-
-	// add the payload from the queue
-	if txPayload != nil {
-		// validate the max payload size
-		if len(txPayload.Data) > Band.MaxPayloadSize[rx1DR].N {
-			// remove the payload from the queue regarding confirmed or not
-			if _, err := clearInProcessTXPayload(ctx.RedisPool, ns.DevEUI); err != nil {
-				return err
-			}
-
-			log.WithFields(log.Fields{
-				"dev_eui":             ns.DevEUI,
-				"data_rate":           rx1DR,
-				"frmpayload_size":     len(txPayload.Data),
-				"max_frmpayload_size": Band.MaxPayloadSize[rx1DR].N,
-			}).Warning("downlink payload max size exceeded")
-			err = ctx.Application.SendNotification(ns.AppEUI, ns.DevEUI, models.ErrorNotificationType, models.ErrorPayload{
-				Reference: txPayload.Reference,
-				DevEUI:    ns.DevEUI,
-				Message:   fmt.Sprintf("downlink payload max size exceeded (dr: %d, allowed: %d, got: %d)", rx1DR, Band.MaxPayloadSize[rx1DR].N, len(txPayload.Data)),
-			})
-			if err != nil {
-				return err
-			}
-		} else {
-			// remove the payload from the queue when not confirmed
-			if !txPayload.Confirmed {
-				if _, err := clearInProcessTXPayload(ctx.RedisPool, ns.DevEUI); err != nil {
-					return err
-				}
-			}
-
-			macPL.FHDR.FCtrl.FPending = remaining
-			if txPayload.Confirmed {
-				phy.MHDR.MType = lorawan.ConfirmedDataDown
-			}
-			macPL.FPort = &txPayload.FPort
-			macPL.FRMPayload = []lorawan.Payload{
-				&lorawan.DataPayload{Bytes: txPayload.Data},
-			}
-		}
-	}
-
-	// when the payload did not pass the validation and there is no ACK set,
-	// there is nothing to send
-	if !macPL.FHDR.FCtrl.ACK && len(macPL.FRMPayload) == 0 {
-		return nil
-	}
-
-	// if there is no payload set, encrypt will just do nothing
-	if err := phy.EncryptFRMPayload(ns.AppSKey); err != nil {
-		return fmt.Errorf("encrypt FRMPayload error: %s", err)
-	}
-
-	if err := phy.SetMIC(ns.NwkSKey); err != nil {
-		return fmt.Errorf("set MIC error: %s", err)
-	}
-
-	txPacket := models.TXPacket{
-		TXInfo: models.TXInfo{
-			MAC:       rxPacket.RXInfo.MAC,
-			Timestamp: rxPacket.RXInfo.Timestamp + uint32(rxDelay/time.Microsecond),
-			Frequency: Band.DownlinkChannels[rx1Channel].Frequency,
-			Power:     Band.DefaultTXPower,
-			DataRate:  Band.DataRates[rx1DR],
-			CodeRate:  rxPacket.RXInfo.CodeRate,
-		},
-		PHYPayload: phy,
-	}
-
-	// window 1
-	if err := ctx.Gateway.SendTXPacket(txPacket); err != nil {
-		return fmt.Errorf("send tx packet (rx window 1) to gateway error: %s", err)
-	}
-
-	// increment the FCntDown when MType != ConfirmedDataDown. In case of
-	// ConfirmedDataDown we increment on ACK.
-	if phy.MHDR.MType != lorawan.ConfirmedDataDown {
-		ns.FCntDown++
-		if err := saveNodeSession(ctx.RedisPool, ns); err != nil {
-			return err
-		}
-	}
-
-	return nil
 }
